moved interface to local | changes extension .ts
