+++ components/incidentError/IncidentErrorPage.tsx
@@ -3,11 +3,14 @@ import Link from "next/link";
 
 //components
 import { Block } from "baseui/block";
-import { IncidentErrorPageProps } from "../../variableTypes";
 import {
   ERROR_PAGE_OVERRIDES,
   MAIN_STYLE_OVERRIDES,
 } from "../incidents/internal/form/overrides/BlockStyles";
+
+interface IncidentErrorPageProps {
+  message: string;
+}
 
 /*Will be displayed if error in processing incident update request or fetching components* */
