+++ types.ts
@@ -1,0 +1,17 @@
+export interface STATUSType {
+  [key: string]: number;
+}
+
+export interface ComponentsJSONObject {
+  name: String;
+  id: String;
+  status: string;
+}
+
+export interface ComponentObject {
+  compName: string;
+  compType: number;
+  id: string;
+  compId: string;
+  selected: boolean;
+}
